+++ packages/wonder-blocks-modal/src/components/modal-header.tsx
@@ -2,26 +2,14 @@ import * as React from "react";
 import {Breadcrumbs} from "@khanacademy/wonder-blocks-breadcrumbs";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {HeadingMedium, LabelSmall} from "@khanacademy/wonder-blocks-typography";
-import {
-    ThemedStylesFn,
-    useScopedTheme,
-    useStyles,
-} from "@khanacademy/wonder-blocks-theming";
-import {
-    ModalDialogThemeContext,
-    ModalDialogThemeContract,
-} from "../themes/themed-modal-dialog";
+import {StyleSheet} from "aphrodite";
+import theme from "../theme";
 
 type Common = {
     /**
      * The main title rendered in larger bold text.
      */
     title: string;
-    /**
-     * Whether to display the "light" version of this component instead, for
-     * use when the item is used on a dark background.
-     */
-    light: boolean;
     /**
      * An id to provide a selector for the title element.
      */
@@ -82,8 +70,6 @@ type Props = Common | WithSubtitle | WithBreadcrumbs;
  * - Add a title (required).
  * - Optionally add a subtitle or breadcrumbs.
  * - We encourage you to add `titleId` (see Accessibility notes).
- * - If the `ModalPanel` has a dark background, make sure to set `light` to
- *   `false`.
  * - If you need to create e2e tests, make sure to pass a `testId` prop and
  *   add a sufix to scope the testId to this component: e.g.
  *   `some-random-id-ModalHeader`. This scope will also be passed to the title
@@ -96,7 +82,6 @@ type Props = Common | WithSubtitle | WithBreadcrumbs;
  *      title="Sidebar using ModalHeader"
  *      subtitle="subtitle"
  *      titleId="uniqueTitleId"
- *      light={false}
  *  />
  * ```
  */
@@ -104,7 +89,6 @@ export default function ModalHeader(props: Props) {
     const {
         // @ts-expect-error [FEI-5019] - TS2339 - Property 'breadcrumbs' does not exist on type 'Readonly<Props> & Readonly<{ children?: ReactNode; }>'.
         breadcrumbs = undefined,
-        light,
         // @ts-expect-error [FEI-5019] - TS2339 - Property 'subtitle' does not exist on type 'Readonly<Props> & Readonly<{ children?: ReactNode; }>'.
         subtitle = undefined,
         testId,
@@ -116,11 +100,8 @@ export default function ModalHeader(props: Props) {
         throw new Error("'subtitle' and 'breadcrumbs' can't be used together");
     }
 
-    const {theme} = useScopedTheme(ModalDialogThemeContext);
-    const styles = useStyles(themedStylesFn, theme);
-
     return (
-        <View style={[styles.header, !light && styles.dark]} testId={testId}>
+        <View style={[styles.header]} testId={testId}>
             {breadcrumbs && (
                 <View style={styles.breadcrumbs}>{breadcrumbs}</View>
             )}
@@ -134,7 +115,7 @@ export default function ModalHeader(props: Props) {
             </HeadingMedium>
             {subtitle && (
                 <LabelSmall
-                    style={light && styles.subtitle}
+                    style={styles.subtitle}
                     testId={testId && `${testId}-subtitle`}
                 >
                     {subtitle}
@@ -151,7 +132,7 @@ export default function ModalHeader(props: Props) {
  */
 const small = "@media (max-width: 767px)";
 
-const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
+const styles = StyleSheet.create({
     header: {
         // TODO(WB-1878): Move this to an `elevation` theme token.
         boxShadow: `0px 1px 0px ${theme.header.color.border}`,
@@ -163,14 +144,9 @@ const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
         position: "relative",
         width: "100%",
 
-        [small]: {
+        [small as any]: {
             paddingInline: theme.header.spacing.paddingInlineSm,
         },
-    },
-
-    dark: {
-        background: theme.root.color.inverse.background,
-        color: theme.root.color.inverse.foreground,
     },
 
     breadcrumbs: {
@@ -181,7 +157,7 @@ const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
     title: {
         // Prevent title from overlapping the close button
         paddingRight: theme.header.spacing.titleGapMd,
-        [small]: {
+        [small as any]: {
             paddingRight: theme.header.spacing.titleGapSm,
         },
     },
@@ -191,7 +167,3 @@ const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
         marginTop: theme.header.spacing.gap,
     },
 });
-
-ModalHeader.defaultProps = {
-    light: true,
-};