+++ packages/wonder-blocks-modal/src/components/modal-panel.tsx
@@ -2,20 +2,13 @@ import * as React from "react";
 import {PropsFor, View} from "@khanacademy/wonder-blocks-core";
 import type {StyleType} from "@khanacademy/wonder-blocks-core";
 
-import {
-    ThemedStylesFn,
-    useScopedTheme,
-    useStyles,
-} from "@khanacademy/wonder-blocks-theming";
-import {actionStyles, focusStyles} from "@khanacademy/wonder-blocks-styles";
+import {focusStyles} from "@khanacademy/wonder-blocks-styles";
+import {StyleSheet} from "aphrodite";
 import ModalContent from "./modal-content";
 import ModalHeader from "./modal-header";
 import ModalFooter from "./modal-footer";
 import CloseButton from "./close-button";
-import {
-    ModalDialogThemeContext,
-    ModalDialogThemeContract,
-} from "../themes/themed-modal-dialog";
+import theme from "../theme";
 
 type Props = {
     /**
@@ -42,11 +35,6 @@ type Props = {
      * become too tall?
      */
     scrollOverflow: boolean;
-    /**
-     * Whether to display the "light" version of this component instead, for
-     * use when the item is used on a dark background.
-     */
-    light: boolean;
     /**
      * Any optional styling to apply to the panel.
      */
@@ -91,7 +79,6 @@ type Props = {
 export default function ModalPanel({
     closeButtonVisible = true,
     scrollOverflow = true,
-    light = true,
     content,
     footer,
     header,
@@ -99,9 +86,6 @@ export default function ModalPanel({
     style,
     testId,
 }: Props) {
-    const {theme} = useScopedTheme(ModalDialogThemeContext);
-    const styles = useStyles(themedStylesFn, theme);
-
     const renderMainContent = React.useCallback((): React.ReactNode => {
         const mainContent = ModalContent.isComponentOf(content) ? (
             (content as React.ReactElement<PropsFor<typeof ModalContent>>)
@@ -122,24 +106,19 @@ export default function ModalPanel({
             // know about things being positioned around it.
             style: [!!footer && styles.hasFooter, mainContent.props.style],
         });
-    }, [content, footer, scrollOverflow, styles.hasFooter]);
+    }, [content, footer, scrollOverflow]);
 
     const mainContent = renderMainContent();
 
-    const isInverse = !light;
-
     return (
         <View
-            style={[styles.wrapper, isInverse && styles.dark, style]}
+            style={[styles.wrapper, style]}
             testId={testId && `${testId}-panel`}
         >
             {closeButtonVisible && (
                 <CloseButton
                     onClick={onClose}
-                    style={[
-                        styles.closeButton,
-                        isInverse && actionStyles.inverse,
-                    ]}
+                    style={[styles.closeButton]}
                     testId={testId && `${testId}-close`}
                 />
             )}
@@ -157,10 +136,9 @@ export default function ModalPanel({
 ModalPanel.defaultProps = {
     closeButtonVisible: true,
     scrollOverflow: true,
-    light: true,
 };
 
-const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
+const styles = StyleSheet.create({
     wrapper: {
         flex: "1 1 auto",
         position: "relative",
@@ -184,14 +162,7 @@ const themedStylesFn: ThemedStylesFn<ModalDialogThemeContract> = (theme) => ({
         // NOTE: IconButton uses :focus-visible, which is not supported for
         // programmatic focus. This is a workaround to make sure the focus
         // outline is visible when this control is focused.
-        ":focus": {
-            ...focusStyles.focus[":focus-visible"],
-        },
-    },
-
-    dark: {
-        background: theme.root.color.inverse.background,
-        color: theme.root.color.inverse.foreground,
+        ":focus": focusStyles.focus[":focus-visible"],
     },
 
     hasFooter: {