+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @khanacademy/wonder-blocks-link
 
+## 9.1.9
+
+### Patch Changes
+
+- Updated dependencies [0f4f771]
+    - @khanacademy/wonder-blocks-tokens@11.1.0
+    - @khanacademy/wonder-blocks-clickable@7.1.9
+    - @khanacademy/wonder-blocks-icon@5.2.3
+
 ## 9.1.8
 
 ### Patch Changes