+++ packages/wonder-blocks-typography/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-typography
 
+## 4.2.4
+
+### Patch Changes
+
+- Updated dependencies [0f4f771]
+    - @khanacademy/wonder-blocks-tokens@11.1.0
+
 ## 4.2.3
 
 ### Patch Changes